+++ src/app/layout.tsx
@@ -1,8 +1,10 @@
 //import { Geist, Geist_Mono } from "next/font/google";
 
 import { BaseCol } from '@/components/layout/base-col'
+import { ThemeProvider } from '@/lib/edb/theme'
 import { Auth0Provider } from '@auth0/nextjs-auth0/client'
 import { Geist } from 'next/font/google'
+import Script from 'next/script'
 import { ReactNode } from 'react'
 import './globals.css'
 const primaryFont = Geist({
@@ -32,6 +34,29 @@ export default function Layout({
 }>) {
   return (
     <html lang="en" className="h-full">
+      <head>
+        {/* This script runs immediately before the page interactive phase */}
+        <Script id="theme-initializer" strategy="beforeInteractive">
+          {`
+            (function() {
+              try {
+                const savedTheme = localStorage.getItem('theme');
+                const systemPrefersDark = window.matchMedia('(prefers-color-scheme: dark)').matches;
+                
+                if (savedTheme === 'dark' || (!savedTheme && systemPrefersDark)) {
+                  document.documentElement.classList.add('dark');
+                } else {
+                  document.documentElement.classList.remove('dark');
+                }
+
+                console.log('Theme initialized')
+              } catch (e) {
+                console.error('Theme initialization failed:', e);
+              }
+            })();
+          `}
+        </Script>
+      </head>
       <body
         className={`${primaryFont.className} antialiased margin-0 h-full min-h-screen flex flex-col bg-body text-base font-normal`}
       >
@@ -39,9 +64,11 @@ export default function Layout({
         {/* <CoreProviders>{children}</CoreProviders> */}
 
         {/* Added for base-ui to render dialogs */}
-        <BaseCol className="root isolate grow">
-          <Auth0Provider>{children}</Auth0Provider>
-        </BaseCol>
+        <ThemeProvider>
+          <BaseCol className="root isolate grow">
+            <Auth0Provider>{children}</Auth0Provider>
+          </BaseCol>
+        </ThemeProvider>
       </body>
     </html>
   )